+++ Assets/TestScript.cs
@@ -1,15 +1,22 @@
 
 using System;
 using System.Collections;
+using System.Reflection;
 using NoodleKit;
 using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.Rendering.Universal;
 using Unity.Mathematics;
+using Unity.VisualScripting;
+using UnityEditor.Experimental.GraphView;
 
 public class TestScript : MonoBehaviour {
-    private void Start() {
-        
+    
+    [Button("Bla", "holy shit", true)]
+    public float IExist;
+
+    public void Bla() {
+        Debug.Log("OMG");
     }
 }