+++ Assets/StarterKit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.2.0]
+### BREAKING CHANGES
+- Made Button Attribute work without UnityEvents. Just give it the method name now.
+- Renamed the old Button Attribute to EventButton.
+
+### Fixes
+- Made Checklist Item handle centered for multi-line items.
+
 ## [1.1.0]
 ### Editor Scripts
 - Added Units Attribute